Improve AFK detection
instead of checking for directional inputs (A/D), the current direction
is now compared to the previous one. So if the input direction does not
change for x ticks, the player is afk. This change was made to ensure
that players that run against a wall and go afk, are detected properly.